security changes to uninstall logic security changes to user permissions

+++ uninstall.php
@@ -11,12 +11,14 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
 
 global $wpdb;
 
-$table = $wpdb->prefix . 'nutrition_short_urls';
-
-// phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
-$wpdb->query("DROP TABLE IF EXISTS {$table}");
+if (get_option('nutrition_labels_delete_data_on_uninstall') === 'yes') {
+  $table = $wpdb->prefix . 'nutrition_short_urls';
+  // phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
+  $wpdb->query("DROP TABLE IF EXISTS {$table}");
+}
 
 delete_option('qr_size');
 delete_option('qr_format');
 delete_option('qr_error_correction');
 delete_option('nutrition_labels_db_version');
+delete_option('nutrition_labels_delete_data_on_uninstall');
